altera path, corrige importações, formata a data, deixa os cards com lotes selecionáveis, valida info do agendamento antes de enviar para o db
